jälkikäteen loppuja, osaa 5 varten

+++ osa4/blogilista/controllers/blogs.js
@@ -1,36 +1,84 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/',  async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs.map(formatBlog))
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1, name: 1})
+    response.json(blogs.map(Blog.format))
 })
 
-blogsRouter.post('/',  async (request, response) => {
+blogsRouter.get('/:id',  async (request, response) => {
     try {
-        const blog = new Blog(request.body)
+        const blog = await Blog.findById(request.params.id)
 
-        if (blog.url === undefined || blog.title === undefined) {
-            return response.status(400).json({ error: 'important information missing' })
+        if (blog) {
+            response.json(Blog.format(blog))
+        } else {
+            response.status(404).json({ error: 'cannot find blog' })
         }
-
-        if (blog.likes === undefined) {
-            blog.likes = 0
-        }
-    
-        const savedBlog = await blog.save()
-        response.json(formatBlog(savedBlog))
     } catch (exception) {
         console.log(exception)
         response.status(500).json({ error: 'something went wrong...' })
     }
 })
 
+blogsRouter.post('/',  async (request, response) => {
+    try {
+        const body = request.body
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        if (body.url === undefined || body.title === undefined) {
+            return response.status(400).json({ error: 'important information missing' })
+        }
+
+        if (body.likes === undefined) {
+            body.likes = 0
+        }
+
+        const user = await User.findById(decodedToken.id)
+
+        const blog = new Blog(body)
+        blog.user = user._id
+    
+        const savedBlog = await blog.save()
+
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
+        response.json(Blog.format(savedBlog))
+    } catch (exception) {
+        if (exception.name === 'JsonWebTokenError' ) {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(500).json({ error: 'something went wrong...' })
+        }
+    }
+})
+
 blogsRouter.delete('/:id',  async (request, response) => {
     try {
-        await Blog.findByIdAndRemove(request.params.id)
-  
-        response.status(204).end()
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const blog = await Blog.findById(request.params.id)
+
+        if (decodedToken.id.toString() === blog.user.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(401).json({ error: 'action not allowed for this user' })
+        }
     } catch (exception) {
         console.log(exception)
         response.status(400).json({ error: 'malformatted id' })
@@ -47,15 +95,5 @@ blogsRouter.put('/:id',  async (request, response) => {
         response.status(400).json({ error: 'malformatted id' })
     }
 })
-
-const formatBlog = (blog) => {
-    return {
-        id: blog._id,
-        title: blog.title,
-        author: blog.author,
-        url: blog.url,
-        likes: blog.likes
-    }
-}
 
 module.exports = blogsRouter
